// apps/backend/src/types/jolpica.api.types.ts
/**
 * Raw Jolpica API types - autogenerated from https://github.com/jolpica/jolpica-f1/blob/main/docs/README.md
 */

export interface JolpicaDriver {
  driverId: string;
  givenName: string;
  familyName: string;
  dateOfBirth: string;
  nationality: string;
  url: string;
}

export interface JolpicaConstructor {
  constructorId: string;
  name: string;
  nationality: string;
  url: string;
}

export interface JolpicaDriverStanding {
  position: string;
  wins: string;
  Driver: JolpicaDriver;
  Constructors: JolpicaConstructor[];
}

export interface JolpicaStandingsList {
  season: string;
  round: string;
  DriverStandings: JolpicaDriverStanding[];
}

export interface JolpicaCircuit {
  circuitName: string;
  Location: {
    country: string;
    locality: string;
  };
}

export interface JolpicaRaceResult {
  Driver: JolpicaDriver;
  Constructor: JolpicaConstructor;
  laps: string;
  Time: {
    time: string;
  };
}

export interface JolpicaRace {
  season: string;
  round: string;
  url: string;
  raceName: string;
  date: string;
  Circuit: JolpicaCircuit;
  Results: JolpicaRaceResult[];
}

export interface JolpicaStandingsResponse {
  MRData: {
    limit: string;
    offset: string;
    total: string;
    StandingsTable: {
      StandingsLists: JolpicaStandingsList[];
    };
  };
}

export interface JolpicaRaceResultsResponse {
  MRData: {
    limit: string;
    offset: string;
    total: string;
    RaceTable: {
      season: string;
      Races: JolpicaRace[];
    };
  };
}